Complete CSS exercise 04
I originally had the height value for the proportional image set to
100%, which did achieve the desired result of the image retaining
its proportions.

After viewing the solution I changed it to auto, since this value is
more appropriate for this situation (and I forgot it existed).
However if I remember correctly the height scales by default to retain
the images proportions without a value being set so the argument
could be made that I could have left it blank.

Best practices though, haha.