Complete the css fix exercise
Followed the exercise to fix up the css for the page to achieve the
correct outcome.

I then checked the solution sheet to compare my original approach to its
contents and reviewed both.

I then added some comments to my style sheet and then set the solution's
approaches into it to try and cement my understanding of the
differences.